@@ -7,11 +7,12 @@
 #
 
 from __future__ import annotations
+from collections.abc import Mapping, Sequence
 import locale
 import os
 import subprocess
 import sys
-from typing import List, Mapping, Optional, Sequence, Tuple
+from typing import Optional
 
 from ._typing import Path
 
@@ -20,7 +21,7 @@ __all__ = ['exec_', 'command', 'which']
 
 
 def exec_(args: Sequence[Path], cwd: Optional[Path] = None, env: Optional[Mapping[str, str]] = None,
-          encoding: Optional[str] = None, errors: str = 'strict') -> Tuple[str, str]:
+          encoding: Optional[str] = None, errors: str = 'strict') -> tuple[str, str]:
     env = dict(env) if env else {}
     env['LC_MESSAGES'] = 'C'
     for k in ('LC_ALL', 'LANG', 'PATH', 'LD_LIBRARY_PATH', 'SystemRoot'):
@@ -46,7 +47,7 @@ def command(name: str, *args: str) -> str:
 
 
 def which(name: str) -> Optional[str]:
-    cands: List[str] = []
+    cands: list[str] = []
     if sys.platform == 'win32':
         cands += (name + ext for ext in os.environ['PATHEXT'].split(os.pathsep))
     cands.append(name)
